requirements.txt
pymupdf
python-dotenv
anthropic
pinecone-client
sentence_transformers
groq